chore: rename service to ConfigurationService
